Move code from registration profile class to utility classes
